Removing Hamming Window use from Audio Encoder
This seems to fix Issue #101. I may be misunderstanding Hamming Window, but it seems like it should be used for parts of the audio that go above the limit and that you want to bring back to within the limits, but still preserve the effect by smoothing it out. The fact that we were applying this to the whole recording made the start and end of the audio file quiet, but the middle the loudest. I honestly don't think we should be changing the recording in any way. Users can and likely should be doing that through external tools, such as Audacity. It has features such as Normalization, which would likely accomplish what this hamming window was trying to do.